Uploading version 17 files
These are the files for version 17 of Patel2020. This version aims to abstract the main program into a class called Boussinesq. 

This class has functions for initializin the Boussinesq parameters, computing the solution of the equation, and plotting the solution graphically as an animation through time.

This class will add functions for defining the water depth, as well as the initial an boundary conditions of the domain.