Added FloorProfile class and fixed bugs
Created FloorProfile class that controls the description an creation of a floor profile. This allows you to set the water depth and store presets for the floor profile.

A bug was fixed where the floor profile was loaded into the water height (h) instead of converting it to a water depth.